fix(affinity): u128 over IPC silently broke every Custom-mask save
User reported every affinity-rule save through the picker failed with
'save affinity rule: error — parse: u128 is not supported'. Same wire
error was silently killing one-shot Custom-mask pins too — serde_json
rejects 128-bit integers by spec, so the entire CpuSelector::Mask(u128)
variant was a serialization dead end whenever it crossed the named pipe.

Sized to u64 (Windows KAFFINITY width — one bit per logical processor
in a single processor group, up to 64). Multi-group machines use
GROUP_AFFINITY, which this selector doesn't model anyway, so u128 was
over-engineering from day one. Touches:

  * crates/core/src/topology.rs — Mask(u128) → Mask(u64), test fixture,
    resolve() shift to 1u64
  * crates/tray/src/main.rs — picker hex parse + display + selector
    construction call sites
  * crates/tray/src/formatters.rs — selector label formatter (no
    width-specific change but recompiled against new variant)

Companion improvement: when 'Save as rule' is used (picker or right-
click Remember), send a single SetAffinityRule { apply_to_live: true }
instead of SetProcessAffinity + SetAffinityRule. The engine's
apply_to_live walk pins each matching live PID AND marks it in
`affinity_rule_applied` so the 2 s re-assert sweep covers it
immediately — closing a ~10 s window where the previous two-IPC pattern
left the freshly-pinned PID unmarked (set_affinity_rule clears the
marker set on every policy mutation), letting a game-overridden pin
slip until the next 10 s background scan re-marked it.

The 'pin sticks under load' Process-Lasso-parity behavior depends on
both fixes together: the rule has to actually save, and the live PID
has to be covered by the re-assert sweep from t=0 rather than t=10s.

+++ crates/core/src/topology.rs
@@ -216,7 +216,7 @@ impl CpuTopology {
                     .collect()
             }
             CpuSelector::Mask(mask) => (0..self.count() as u32)
-                .filter(|i| mask & (1u128 << i) != 0)
+                .filter(|i| mask & (1u64 << i) != 0)
                 .collect(),
         }
     }
@@ -246,7 +246,14 @@ pub enum CpuSelector {
     /// Top N cores by CPPC perf rank (with SMT siblings de-prioritised).
     TopRanked(u32),
     /// Explicit bitmask over logical processor indices. Last resort / legacy.
-    Mask(u128),
+    ///
+    /// `u64` matches Windows' `KAFFINITY` width — one bit per logical processor
+    /// in the calling thread's group, up to 64. Multi-group machines use
+    /// `GROUP_AFFINITY` instead and aren't covered by this selector. Sized as
+    /// `u64` (not `u128`) so the wire format JSON can serialize it — serde_json
+    /// rejects 128-bit integers by spec, and that rejection silently broke
+    /// every Custom-mask save through the IPC pipe before this fix.
+    Mask(u64),
 }
 
 #[cfg(test)]
@@ -336,7 +343,7 @@ mod tests {
     fn resolve_mask_picks_only_set_bits() {
         let topo = x3d_dual_ccd();
         // Bits 0, 2, 4, 6 → first four even logical indices.
-        let mask: u128 = 0b0101_0101;
+        let mask: u64 = 0b0101_0101;
         let resolved = topo.resolve(&CpuSelector::Mask(mask));
         assert_eq!(resolved, vec![0, 2, 4, 6]);
     }
